Fix validation for answers that include percent signs
Summary:
Wow, this is really bad! Our mobile input mechanisms (MyScript, and the custom math keypad) operate
over TeX rather than plaintext. When users entered percent-signs, they were represented in the TeX
as '\%', rather than '%' (since the percent-sign is a comment character in TeX). As such,
validation was //always// failing when an answer included a percent-sign, since the TeX would say,
e.g., '4\%', and `input-number` would expect '4%' exactly. This has probably been happening for the
mobile apps since we first shipped them, and for mobile web since we shipped the custom keypad.

We had to solve this same problem, in the past, for fractions, which TeX represents as `\dfrac` or
`\frac`, but `input-number` wants to see as simple '/' characters. To fix the backslashing, I've
augmented the transform that we run over our inputs before passing them to `input-number` and
`numeric-input` to do a simple search-and-replace.

On Asana @ https://app.asana.com/0/257584043292808/284212465920286.

Test Plan:
- Substate into webapp.
- Run: `make serve`.
- Run: `make watch` (in Perseus).
- Open an exercise that requires an answer with a percent-sign (`http://localhost:8080/math/statistics-probability/significance-tests-one-sample/idea-of-significance-tests/e/hypothesis-testing-with-simulations`).
- Take all the hints; plug in the right answer.
    - Verify that it's marked as correct.
- (You can also log in `input-number.jsx#validate` to ensure that the input was transformed from something that contained '\%' to something that contained only '%').

Reviewers: kevinb, benkomalo, michaelpolyak, johnsullivan

Reviewed By: benkomalo

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D33917