Add some keys to the added spans in passages.
Summary:
We've recently been getting some react warnings about keys in spans in
passage widgets. With the React 15 upgrade, we added some extra explicit spans
in passages, but didn't add keys to them. This adds appropriate keys to the new
spans (I think it was the line number spans that were actually causing the
problem, but it's so easy to add keys in simple-markdown that I figured I would
do it there too while I was at it)

Test Plan:
 - Visit [this multi-item with a
   passage](http://localhost:9000/?multirenderer#content=%7B%22_multi%22%3A%7B%22sharedContext%22%3A%7B%22__type%22%3A%22content%22%2C%22content%22%3A%22%5B%5B%E2%98%83%20passage%201%5D%5D%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%22passage%201%22%3A%7B%22options%22%3A%7B%22static%22%3Afalse%2C%22passageTitle%22%3A%22%22%2C%22passageText%22%3A%22Did%20music%20and%20human%20language%20originate%20separately%20or%20together%3F%20Both%20systems%20use%20intonation%20and%20rhythm%20to%20communicate%20emotions.%22%2C%22footnotes%22%3A%22%22%2C%22showLineNumbers%22%3Atrue%7D%2C%22type%22%3A%22passage%22%2C%22version%22%3A%7B%22major%22%3A0%2C%22minor%22%3A0%7D%2C%22graded%22%3Atrue%2C%22alignment%22%3A%22default%22%2C%22static%22%3Afalse%7D%7D%7D%2C%22questions%22%3A%5B%5D%2C%22hints%22%3A%5B%5D%7D%7D)
 - See that there are no react warnings about keys

Reviewers: mattdunnrankin

Reviewed By: mattdunnrankin

Subscribers: lauren, jared

Differential Revision: https://phabricator.khanacademy.org/D34075