[multirenderer] Make multirenderer <Renderer>s react to apiOptions changes
Summary:
Before this change, if new `apiOptions` were passed in to the
multirenderer through props, those changes wouldn't get propogated to the
`<Renderer>`s in the tree, because we stored the value of the props at the
moment when we generate the `<Renderer>`s.

One solution would be to re-generate the whole renderer tree when the
`apiOptions` prop (or any other prop) was changed. This seems excessive.
Instead, this change makes it so that each time we render, we make a
`<Renderer>` with the current extra props.

Test Plan:
 - Visit http://localhost:9000/?multirenderer
 - `make build`
 - Visit a question in test prep and change some apiOptions without changing
   the content of the question
 - See that the new apiOptions are applied.

Reviewers: mattdunnrankin

Reviewed By: mattdunnrankin

Subscribers: lauren, jared, #sat

Differential Revision: https://phabricator.khanacademy.org/D34241