Add question serializedState to the score.
Summary: At the moment we only save the guess data for an exercise attempt, however the serializedState for a question seems to be far more useful.

Test Plan: I'm not sure there's much to test here but I have a follow-up diff in `webapp` which includes this change and verifies that it works.

Reviewers: emily, alex

Reviewed By: emily

Subscribers: sean, jangmi, dhruv, jared

Differential Revision: https://phabricator.khanacademy.org/D34358