[multirenderer] Make the MultiRenderer handle updating serialized state.
Summary:
This adds `serializedState` and `onSerializedStateUpdated` props to
the `MultiRenderer` to allow for clients to subscribe to serialized state
changes and declaratively specify the current serialized state.

In addition, in an attempt to solve the problem that we can't fetch serialized
state from not-currently-mounted content renderers, we allow for pulling data
from the previous serialized state instead of just returning `null` or a
placeholder.

I left the old `getSerializedState()` and `restoreSerializedState()` functions,
but maybe we should mark them internal since they're pretty janky?

Test Plan:
 - `make build`
 - Use these features in webapp somewhere
 - see that they work?

Reviewers: mattdunnrankin

Reviewed By: mattdunnrankin

Subscribers: lauren, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D34463