[highlighting] Merge overlapping highlights
Summary:
Previously, it was possible to add a highlight that overlapped an existing highlight, which would confusingly cause the overlapping words to be double-highlighted. This version instead deletes the old highlights, and merges their contents into the new highlight.

Depends on D35100, the first draft of the highlighting refactor.

Test Plan:
Create highlights, and confirm that they merge when expected. In particular, check for:

1. New highlight is fully contained by an existing highlight.
2. New highlight starts in the middle of one highlight, fully contains multiple additional highlight, and ends in the middle of a final highlight.

Confirm that the highlights merged, by deleting the new highlight and confirming that it's all deleted at once.

[[ https://app.asana.com/0/277557989281705/319354409889328 | Asana task ]]

Reviewers: amy, emily, nickbreen

Reviewed By: nickbreen

Subscribers: lauren, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D35118