[cp mcr] Make mobile content library rationales look good
Summary:
The goal of this diff is to implement the current designs for the
content library rationales (otherwise known as "clues" in the code), while not
affecting the visuals of multiple choice answers in any other way.

In particular, solving this involved moving a bunch of the padding around
multiple choice options down the tree so that we can achieve the correct
padding around both the description and rationale, when they are shown.

I'm pretty sure the only visual change that this diff makes to existing
multiple choice questions is removing the `opacity: 0.5` we used to add when
you were done answering a question. @nefaurk I'd love guidance on whether it's
okay to ship that!

Screenshot:
{F686562}

Test Plan:
 - Visit
   http://localhost:8081/science/physics/one-dimensional-motion/acceleration-tutorial/e/rationales-test
   in iPhone 6 device emulation.
 - See that the visuals of the multiple choice questions hasn't changed.
 - Get an answer correct
 - See visually correct rationales!
 - Get questions correct on >1280px pages, see that rationales show up and
 - Get questions correct on <1280px pages, see that rationales show up and look
   not-awful.
 - Visit http://localhost:8081/mission/sat
 - Practice a math task, see that the multiple choice answers look the same.
   Answer a question, see rationales show up and see that they look the same
   too.

Reviewers: alex, johnsullivan, michaelpolyak

Reviewed By: michaelpolyak

Subscribers: nefaurk, lauren, jared, #contentplatform

Differential Revision: https://phabricator.khanacademy.org/D35202