[highlighting] Fake hover and click events for highlights
Summary:
In order to allow the highlighted content to receive the browser's native hover and click events, we render the highlights behind the content. But, then, if we want to know whether a highlight is being hovered how do we do it?

Answer: we reimplement hover ourselves! Nope, not joking. The alternative is to mess with the content's DOM, and this hack, while indeed quite hacky, is wonderfully self-contained.

Depends on D35100, the first draft of the highlighting refactor.

(Like the parent revision, this draft of this code is missing lots of documentation - I rushed through it to prove that it's feasible. Now that the riskiest stuff is done, though, I'm gonna revisit the parent and add more docs :D)

[[ https://app.asana.com/0/277557989281705/319354409889332 | Asana task ]]

Test Plan:
{F684568}
Create some highlights, then hover over one and confirm that its background changes.

Confirm that, when hovering over a highlight, your mouse cursor is still the text cursor, indicating that the content is in the foreground.

Finally, confirm that you can still click a highlight to remove it. (This isn't the final behavior for highlight removal; it's just an important feature that I didn't want to regress on, so hacked together here with global click listeners :P)

Reviewers: amy, emily, nickbreen, joshua

Reviewed By: amy

Subscribers: lauren, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D35126