[highlighting] "Remove highlight" tooltip
Summary:
Previously, to remove a highlight, you would click on it. Now, we instead show a "Remove highlight" tooltip on hover, that, when clicked, removes the highlight.

This change also extracts some of the tooltip-positioning logic out of `SelectionTracker` and into `HighlightTooltip`. This also removes the need for `SelectionTracker`'s rather particular `shouldComponentUpdate` function, which existed because the hidden-tooltip state made it difficult to distinguish between state updates that required or didn't require a re-render. But, now, most of the actual work is in `HighlightTooltip`, which is a `PureComponent`, so everything is well-cached without us having to intervene :)

Depends on D35298.

[[ https://app.asana.com/0/277557989281705/318794715865769 | Asana task ]]

Test Plan:
{F689426}

Note that we still suffer from the same tooltip wrapping problem :'(

Reviewers: amy, nickbreen

Reviewed By: amy

Subscribers: joshua, lauren, emily, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D35302