[highlighting] Read-only highlights in review mode
Summary:
This change adds an `editable` prop to highlighting components that enable adding/removing highlights. When `editable={false}`, adding and removing highlights is no longer allowed.

When a `Passage` widget is in review mode, editable is set to `false`.

Test Plan:
http://localhost:8081/prep-for/lsat/task/TASK_ID/question/1

Add some highlights, and confirm that the "Remove highlight" tooltip appears as expected.

Answer the question, and confirm that the "Add highlight" tooltip no longer appears on text selection, and "Remove highlight" no longer appears on highlight hover.

Navigate to a different question, and confirm that highlighting works again :)

Additionally, use the React DevTools to confirm that mousemove events are not tracked when in read-only mode, as an optimization.

Reviewers: amy, nickbreen

Reviewed By: nickbreen

Subscribers: joshua, lauren, emily, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D35328