[cp mcr] Remove the onePerLine option in radio widgets
Summary:
There used to be an option in radio widgets which let you choose for
the options to be displayed side-by-side instead of vertically. This was mostly
to save the space taken up by displaying the options vertically. XOM deprecated
this option and removed the ability for content creators to select the
horizontal layout, but never actually removed the ability to show things in a
horizontal way. Rationales look bad when shown below horizontally-layed-out
options, so this diff is removing the option!

Before landing this, we probably should find a bunch of examples of the
horizontally-layed-out items and make sure that they look good when layed out
vertically. In particular, items with images in the choices might look
particularly bad vertically, and we can decide how much we care about that.

Screenshots:

Before:
{F692868}
After:
{F692866}

Before:
{F692869}
After:
{F692870}

Test Plan:
 - Visit [this rendered item with onePerLine set to `false`](http://localhost:9000/?renderer#content=%7B%22question%22%3A%7B%22content%22%3A%22%5B%5B%E2%98%83%20radio%201%5D%5D%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%22radio%201%22%3A%7B%22type%22%3A%22radio%22%2C%22alignment%22%3A%22default%22%2C%22static%22%3Afalse%2C%22graded%22%3Atrue%2C%22options%22%3A%7B%22choices%22%3A%5B%7B%22correct%22%3Atrue%2C%22content%22%3A%22a%22%7D%2C%7B%22correct%22%3Afalse%2C%22content%22%3A%22b%22%7D%2C%7B%22isNoneOfTheAbove%22%3Afalse%2C%22content%22%3A%22c%22%7D%5D%2C%22randomize%22%3Afalse%2C%22multipleSelect%22%3Afalse%2C%22displayCount%22%3Anull%2C%22hasNoneOfTheAbove%22%3Afalse%2C%22deselectEnabled%22%3Afalse%2C%22onePerLine%22%3Afalse%7D%2C%22version%22%3A%7B%22major%22%3A1%2C%22minor%22%3A0%7D%7D%7D%7D%2C%22answerArea%22%3A%7B%22calculator%22%3Afalse%2C%22chi2Table%22%3Afalse%2C%22periodicTable%22%3Afalse%2C%22tTable%22%3Afalse%2C%22zTable%22%3Afalse%7D%2C%22itemDataVersion%22%3A%7B%22major%22%3A0%2C%22minor%22%3A1%7D%2C%22hints%22%3A%5B%5D%7D)
 - See that the options are layed out vertically

Reviewers: kevinb, michaelpolyak, johnsullivan, alex

Reviewed By: johnsullivan

Subscribers: nefaurk, lauren, jared, #contentplatform

Differential Revision: https://phabricator.khanacademy.org/D35332