Update eslint-disable messages in JS files
Summary:
We'd like to update the eslint rules but before that let's update
all the messages to match the new format so that further changes
will reflect changes required by changes in the lint rules.

The script that made these changes was added to the khan-lint repo in D35276.

Test Plan:
- ka-lint src

Auditors: mattdunnrankin