Disable react/sort-comp lint rule in files where props comes after lifecycle methods
Summary:
We'd like to update khan-linter to use
        "react/sort-comp": [2, {
            "order": [
                "type-annotations",
                "static-methods",
                "lifecycle",
                "everything-else",
                "render"
            ]
        }],

so that we can put flow props before the lifecycle methods.

Before we can do that we need to disable sort-comp in all files
where it would break with this change.

I update these files in this diff by running:
../devtools/khan-linter/update-eslint-disable-lines.js src

Test Plan: arc lint

Reviewers: mattdunnrankin, john, nickbreen

Reviewed By: mattdunnrankin

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D35436