[highlighting] Hide "Remove highlight" tooltips when "Add highlight" is visible
Summary:
In particular, we refactor `SelectionTracker` to instead be a wrapper component that exposes the current selection and proposed highlight, if any, to the entire highlighting UI. The highlighting UI then renders an "Add highlight" tooltip (which used to be `SelectionTracker`'s job), and sets the existing highlights to not be in editable mode until the selection is cleared.

Note that clicking typically clears a selection, so, if you have "Add highlight" up, but then //click// on an existing highlight, the "Add highlight" tooltip will be dismissed, and the "Remove highlight" tooltip will appear :)

Depends on D35623, because it changed some parts of the `SelectionTracker` API.

Asana: https://app.asana.com/0/329800276300868/328741675538357

Test Plan: Make some highlights! Confirm that "Add highlight" and "Remove highlight" still generally behave as expected, but that "Remove highlight" no longer appears when "Add highlight" is present.

Reviewers: amy, nickbreen

Reviewed By: nickbreen

Subscribers: joshua, lauren, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D35630