[highlighting] Extract highlight-set-rendering into HighlightSetRenderer
Summary:
A lot of logic for rendering multiple highlights was starting to build up in `HighlightingUI`. This change extracts it into a single component.

This is not only a modularity win, but a perf win: now, when the "Add highlight" tooltip is visible and existing highlights therefore aren't editable, we temporarily remove the mousemove listeners until the "Add highlight" tooltip is dismissed. This behavior happens pretty straightforwardly, by `HighlightSetRenderer` just watching its `editable` prop :)

Depends on D35630, which motivated this refactor.
Depends on D35624, which created the `DOMHighlightSet` type.

Test Plan:
Test basic behavior, confirm no obvious regressions.

Specifically, confirm that mousemove listeners are restored, by the following process:
1. Add a highlight.
2. Open the highlight in React DevTools. Confirm that, as you move your mouse, its `mouseClientPosition` prop changes.
3. Create a selection. Confirm that, as you move your mouse, the `mouseClientPosition` prop no longer updates, and remains `null`.
4. Dismiss the selection. Confirm that, as you move your mouse, `mouseClientPosition` now changes, as it did before.

Reviewers: amy, nickbreen

Reviewed By: amy

Subscribers: joshua, lauren, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D35632