[highlighting] Only show "Add highlight" for valid highlights
Summary:
Previously, if you selected non-highlightable text (like the space between two words), we showed you the Add Highlight tooltip, but, when clicked, no highlight was added. Now, the tooltip only appears when you select a valid range of highlightable text.

This change includes a bit of a refactor, because the find-the-first-and-last-words process used to be baked into serialization—but now we need it earlier, and it seems strange to perform it twice, and strange for the type system to not encode anything about the validation process.

So, I made the following changes:

* Decoupled the process of //building// a highlight from the process of //adding// it to the current set.
  * The UI now calls `buildHighlight` to turn the `DOMRange` into a `DOMHighlight` object, and receives `null` if the `DOMRange` is not currently highlightable. It caches the `DOMHighlight` in state.
  * If the highlight is not `null`, it shows the "Add highlight" tooltip.
  * When the tooltip is clicked, the UI calls `onAddHighlight`, with the cached `DOMHighlight` as a parameter.
* `DOMHighlight` now includes `firstWordIndex` and `lastWordIndex`.
  * This caches the word indexes, helping us avoid naively converting the `DOMRange` into word indexes for validation, and then doing it all over again when actually adding the highlight. This also trivializes serialization, too!
  * This also enforces in the type system that you can only add highlights whose ranges have been validated, because you can only //have// a highlight object if its range was validated :)

I also have some outstanding TODOs around changing how highlight keys work, because it's problematic for `DOMHighlight`'s key to be created before we actually insert it into the list. I noticed some performance issues, too, in that many events trigger React updates but shouldn't—will investigate!

Test Plan:
```
make test
```

Confirm that selecting the space between words no longer yields "Add highlight", but selecting ranges of actual words does.

Confirm that merging highlight ranges still works as expected.

Reviewers: amy, nickbreen

Reviewed By: amy

Subscribers: joshua, lauren, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D35623