[highlighting] Disable highlighting until Aphrodite styles are applied
Summary:
We had a bug where, on the intial render of a Perseus `Passage` in the Test Prep app, any highlights initially in state would be measured before Aphrodite styles were applied, and not re-measured after, leading to obviously bad visual bugs:

{F704102}

This change disabled highlighting until Aphrodite styles are applied.

This sledgehammer of disabled/enabled might not be appropriate for all future use cases, though, in which case we'll need an API to re-measure already-rendered highlights. That more general API is actually what I tried first, but it produced a visible "jump" on initial render between the bad positions and the good positions, and I concluded that we needed to never render highlights in the bad position. So, I went with this approach, which doesn't apply highlights //at all// until the good position is ready :)

Test Plan:
1. Open a passage item in Test Prep.
2. Add a highlight.
3. Wait a few seconds for task state to persist to the server, then refresh.
4. Confirm that the highlight is immediately in the correct position.

Asana: https://app.asana.com/0/329800276300868/335886663293745

Reviewers: amy, nickbreen

Reviewed By: amy

Subscribers: joshua, lauren, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D35655