[cp support] Don't render JIPT labels when the JIPT data didn't load
Summary:
For web+graphie images, we store the label data in a separate file
from the image on S3. To translate these labels, we make a copy of the file
with the translated strings in them. Then, on non-english khan academy, we
download the label file with the labels translated into the current language.
If that file doesn't exist, we fall back to downloading the untranslated
labels.

When doing JIPT (just-in-place translations), we download label files for the
fake `en-pt` language in which the strings are "translated" into unique IDs
that the JIPT script is able to find on the page. In order for the script to
find them, we must not render the labels as math, so in the code we always
rendered the labels as text when doing JIPT.

There's was problem then when we try to load JIPT label data and it fails.
We fall back to untranslated labels, but we still follow the JIPT path and
always render labels as text. This is a problem whe, for instance, a translator
generates a new image with non-english math labels and replaces the old
untranslated graphie image with the new one. There isn't a JIPT translation of
the newly generated image, but we don't render the labels as math even though
they're not JIPT tags.

So, this fixes that problem by keeping track of whether we're using
untranslated or localized label data. Then, we only do the special JIPT
rendering if we're on a JIPT translation page *AND* we're using the "localized"
jipt labels. Otherwise, we do the normal rendering.

Asana task: https://app.asana.com/0/265849773833386/328834014973631

Test Plan:
 - In src/renderer-demo.jsx, add a line like `window.KA = {language: "en-pt"}`
   to emulate a JIPT translation page.
 - Visit [this rendered item with an image with a mathy `1\\text{厘米}` label.
 - See that the image renders the label as math (which we can tell because we
   don't see the `\text{` in the output)

{F704337}

Reviewers: aasmund

Reviewed By: aasmund

Subscribers: tom, lauren, jared, #contentplatform

Differential Revision: https://phabricator.khanacademy.org/D35669