[highlighting] Hide "Add highlight" tooltip if it's beyond the content
Summary:
One of the cool features of the highlighting tool is that, even if your selection extends beyond the boundaries of the content (e.g. you start your highlight in the non-highlightable passage title or directions), we'll still give you the ability to add the highlight, and we automatically trim it to just the valid content words.

However, we didn't apply any validation to the "Add highlight" tooltip, which would appear at the focus of the selection - regardless of whether it was part of the content.

{F704833}

This change explicitly validates whether the selection focus is part of the content, and only shows the tooltip if so.

Asana: https://app.asana.com/0/329800276300868/340038383113611

Test Plan:
Select text that includes some content and some non-content, and whose focus is not in the content. Confirm that no "Add highlight" tooltip is shown.

Select text that includes some content and some non-content, and whose focus //is// in the content. Confirm that the "Add highlight" tooltip is shown.

Confirm that the previous behavior for hiding the "Remove highlight" tooltip behavior is preserved. In particular, notice the subtle case where, as you're mouse-selecting into text that overlaps another highlight, the "Remove highlight" tooltip is hidden, even though the "Add highlight" tooltip is not yet visible.

Confirm that highlight merging still works as expected (because we hoisted the `rangesIntersect` function out of the `mergeRanges` function).

Reviewers: amy, nickbreen

Reviewed By: nickbreen

Subscribers: lauren, jared, joshua, #gtp

Differential Revision: https://phabricator.khanacademy.org/D35680